remove some of the coordinate hacks and let things be backwards till i figure it out
